files/requirements.txt
flask>=2.3.0
torch>=2.0.0
transformers>=4.40.0
pandas>=2.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
accelerate>=0.27.0
python-dotenv>=1.0.0
